+++ ConvertMetricUnits.Core/Repository/Interfaces/ITemparatureRepository.cs
@@ -2,6 +2,6 @@
 {
     public interface ITemparatureRepository
     {
-        double ConvertTemparature(string from, string to, int amount);
+        double ConvertTemparature(string from, string to, double amount);
     }
 }
