+++ ConvertMetricUnits.Core/Repository/LengthRepository.cs
@@ -20,7 +20,7 @@ namespace ConvertMetricUnits.Core.Repository
             _cache = cache;
         }
 
-        public double ConvertLengthAsync(string from, string to, int amount)
+        public double ConvertLengthAsync(string from, string to, double amount)
         {
             var converstionName = string.Concat(from, " to ", to);
             var parameter = DapperParameter.BuildDapperParameters(converstionName);
